PDF page generation (#35)
* First pass at crawling course_files for PDF's and adding pages for them to the course markdown

* First pass at crawling course_files for PDF's and adding pages for them to the course markdown

* First pass at crawling course_files for PDF's and adding pages for them to the course markdown

* fmt

* set type to "file" and layout to "pdf" for PDF pages

* Set type to "courses" for now on pdf pages, because of the way hugo handles template inheritance.

* sections with files should be index.md and not _index.md

* Moved from using the ref shortcode to a custom getpage shortcode to support sections, and using only the PDF's UID for the markdown filename.

* First pass at crawling course_files for PDF's and adding pages for them to the course markdown

* First pass at crawling course_files for PDF's and adding pages for them to the course markdown

* First pass at crawling course_files for PDF's and adding pages for them to the course markdown

* fmt

* set type to "file" and layout to "pdf" for PDF pages

* Set type to "courses" for now on pdf pages, because of the way hugo handles template inheritance.

* sections with files should be index.md and not _index.md

* Moved from using the ref shortcode to a custom getpage shortcode to support sections, and using only the PDF's UID for the markdown filename.

* Remove errant console.log

* refactor out static file writing code into a separate function

* Use pdf filename for markdown filename, but pass through the UID for naming the Javascript functions

* rewrite pdf resolveuid links to their hugo pages by shortcode

* fmt

* Add some asserts to ensure markdown files are created for each section's PDF files.

* optimize pdfFiles map return with object literal

+++ src/file_operations.js
@@ -79,19 +79,36 @@ const writeMarkdownFilesRecursive = (destination, markdownData) => {
     For a given course identifier string and array of objects with properties
     name and data, write Hugo markdown files
     */
-  for (const file of markdownData) {
-    const filePath = path.join(destination, file["name"])
-    const dirPath = path.dirname(filePath)
-    if (!directoryExists(dirPath)) {
-      fs.mkdirSync(dirPath, { recursive: true })
+  for (const section of markdownData) {
+    const sectionPath = path.join(destination, section["name"])
+    const sectionDirPath = path.dirname(sectionPath)
+    if (!directoryExists(sectionDirPath)) {
+      fs.mkdirSync(sectionDirPath, { recursive: true })
     }
-    if (fs.existsSync(filePath)) {
-      fs.unlinkSync(filePath)
+    if (fs.existsSync(sectionPath)) {
+      fs.unlinkSync(sectionPath)
     }
-    fs.writeFileSync(filePath, file["data"])
-    if (file.hasOwnProperty("children")) {
-      writeMarkdownFilesRecursive(destination, file["children"])
+    fs.writeFileSync(sectionPath, section["data"])
+    writeSectionFiles(section, destination)
+    if (section.hasOwnProperty("children")) {
+      writeMarkdownFilesRecursive(destination, section["children"])
     }
+  }
+}
+
+const writeSectionFiles = (section, destination) => {
+  if (section.hasOwnProperty("files")) {
+    section["files"].forEach(file => {
+      const filePath = path.join(destination, file["name"])
+      const fileDirPath = path.dirname(filePath)
+      if (!directoryExists(fileDirPath)) {
+        fs.mkdirSync(fileDirPath, { recursive: true })
+      }
+      if (fs.existsSync(filePath)) {
+        fs.unlinkSync(filePath)
+      }
+      fs.writeFileSync(filePath, file["data"])
+    })
   }
 }
